feat: Complete component refactoring and optimization
🚀 Major Architecture Improvements:
- Extracted App.jsx from 2700+ lines to 200 lines (93% reduction)
- Created modular component structure with proper separation of concerns
- Organized components into logical directories (sections/, background/, hooks/, utils/)

✨ New Components Created:
- 📁 hooks/: useExperienceCalculator, useThrottledScroll
- 🎨 background/: AnimatedParticles, FloatingElements
- 🧭 Navigation: Complete navigation with mobile menu
- 📄 sections/: HeroSection, AboutSection, ExperienceSection, SkillsSection, ProjectsSection, PersonalProjectsSection, TechnologiesSection, ContactSection, Footer

⚡ Performance & Features:
- Added Google Analytics integration with privacy controls
- Maintained all existing functionality and animations
- Improved maintainability and scalability
- Enhanced code organization for team collaboration
- Optimized bundle size and loading performance

🎯 Benefits:
- Better maintainability with isolated components
- Easier testing and debugging capabilities
- Future-ready architecture for adding new features
- Clean separation between business logic and UI components
- Comprehensive analytics tracking for user insights

✅ All tests passing, build successful, ready for production
