+++ src/components/background/index.js
@@ -1,0 +1,2 @@
+export { default as AnimatedParticles } from './AnimatedParticles.jsx';
+export { default as FloatingElements } from './FloatingElements.jsx';
